ai/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.2.0
Pillow>=9.0.0
opencv-python>=4.7.0
tqdm>=4.64.0
tensorboard>=2.12.0
pyyaml>=6.0
albumentations>=1.3.0
grad-cam>=1.4.0
jupyter>=1.0.0
ipykernel>=6.0.0
plotly>=5.0.0